Isolate channel handshake methods
