+++ crates/relayer-cosmos/src/impls/chain.rs
@@ -6,7 +6,7 @@ use ibc_relayer::chain::counterparty::counterparty_chain_from_channel;
 use ibc_relayer::chain::endpoint::ChainStatus;
 use ibc_relayer::chain::handle::ChainHandle;
 use ibc_relayer::chain::requests::{
-    IncludeProof, PageRequest, Qualified, QueryChannelRequest, QueryClientStateRequest,
+    IncludeProof, PageRequest, Qualified, QueryClientStateRequest,
     QueryConsensusStateHeightsRequest, QueryConsensusStateRequest, QueryHeight,
     QueryUnreceivedPacketsRequest,
 };
@@ -34,15 +34,8 @@ use ibc_relayer_types::core::ics02_client::events::CLIENT_ID_ATTRIBUTE_KEY;
 use ibc_relayer_types::core::ics02_client::msgs::create_client::MsgCreateClient;
 use ibc_relayer_types::core::ics03_connection::connection::ConnectionEnd;
 use ibc_relayer_types::core::ics03_connection::version::Version as ConnectionVersion;
-use ibc_relayer_types::core::ics04_channel::channel::{
-    ChannelEnd, Counterparty as ChannelCounterparty, State,
-};
 use ibc_relayer_types::core::ics04_channel::events::{SendPacket, WriteAcknowledgement};
 use ibc_relayer_types::core::ics04_channel::msgs::acknowledgement::MsgAcknowledgement;
-use ibc_relayer_types::core::ics04_channel::msgs::chan_open_ack::MsgChannelOpenAck;
-use ibc_relayer_types::core::ics04_channel::msgs::chan_open_confirm::MsgChannelOpenConfirm;
-use ibc_relayer_types::core::ics04_channel::msgs::chan_open_init::MsgChannelOpenInit;
-use ibc_relayer_types::core::ics04_channel::msgs::chan_open_try::MsgChannelOpenTry;
 use ibc_relayer_types::core::ics04_channel::msgs::recv_packet::MsgRecvPacket;
 use ibc_relayer_types::core::ics04_channel::msgs::timeout::MsgTimeout;
 use ibc_relayer_types::core::ics04_channel::packet::Packet;
@@ -61,6 +54,12 @@ use prost::Message as _;
 use tendermint::abci::Event as AbciEvent;
 
 use crate::contexts::chain::CosmosChain;
+use crate::methods::channel::{
+    build_channel_open_ack_message, build_channel_open_ack_payload,
+    build_channel_open_confirm_message, build_channel_open_confirm_payload,
+    build_channel_open_init_message, build_channel_open_try_message,
+    build_channel_open_try_payload,
+};
 use crate::methods::connection::{
     build_connection_open_ack_message, build_connection_open_ack_payload,
     build_connection_open_confirm_message, build_connection_open_confirm_payload,
@@ -1042,43 +1041,8 @@ where
         height: &Height,
         port_id: &PortId,
         channel_id: &ChannelId,
-    ) -> Result<CosmosChannelOpenTryPayload, Self::Error> {
-        let height = *height;
-        let port_id = port_id.clone();
-        let channel_id = channel_id.clone();
-        let chain_handle = self.handle.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let (channel_end, _) = chain_handle
-                    .query_channel(
-                        QueryChannelRequest {
-                            port_id: port_id.clone(),
-                            channel_id: channel_id.clone(),
-                            height: QueryHeight::Latest,
-                        },
-                        IncludeProof::No,
-                    )
-                    .map_err(BaseError::relayer)?;
-
-                let proofs = chain_handle
-                    .build_channel_proofs(&port_id, &channel_id, height)
-                    .map_err(BaseError::relayer)?;
-
-                let payload = CosmosChannelOpenTryPayload {
-                    previous_channel_id: channel_end.counterparty().channel_id.clone(),
-                    proofs,
-                    ordering: channel_end.ordering,
-                    connection_hops: channel_end.connection_hops,
-                    version: channel_end.version,
-                };
-
-                Ok(payload)
-            })
-            .await
-            .map_err(BaseError::join)?
+    ) -> Result<CosmosChannelOpenTryPayload, Error> {
+        build_channel_open_try_payload(self, height, port_id, channel_id).await
     }
 
     async fn build_channel_open_ack_payload(
@@ -1086,40 +1050,8 @@ where
         height: &Height,
         port_id: &PortId,
         channel_id: &ChannelId,
-    ) -> Result<CosmosChannelOpenAckPayload, Self::Error> {
-        let height = *height;
-        let port_id = port_id.clone();
-        let channel_id = channel_id.clone();
-        let chain_handle = self.handle.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let (channel_end, _) = chain_handle
-                    .query_channel(
-                        QueryChannelRequest {
-                            port_id: port_id.clone(),
-                            channel_id: channel_id.clone(),
-                            height: QueryHeight::Latest,
-                        },
-                        IncludeProof::No,
-                    )
-                    .map_err(BaseError::relayer)?;
-
-                let proofs = chain_handle
-                    .build_channel_proofs(&port_id, &channel_id, height)
-                    .map_err(BaseError::relayer)?;
-
-                let payload = CosmosChannelOpenAckPayload {
-                    proofs,
-                    version: channel_end.version,
-                };
-
-                Ok(payload)
-            })
-            .await
-            .map_err(BaseError::join)?
+    ) -> Result<CosmosChannelOpenAckPayload, Error> {
+        build_channel_open_ack_payload(self, height, port_id, channel_id).await
     }
 
     async fn build_channel_open_confirm_payload(
@@ -1127,26 +1059,8 @@ where
         height: &Height,
         port_id: &PortId,
         channel_id: &ChannelId,
-    ) -> Result<CosmosChannelOpenConfirmPayload, Self::Error> {
-        let height = *height;
-        let port_id = port_id.clone();
-        let channel_id = channel_id.clone();
-        let chain_handle = self.handle.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let proofs = chain_handle
-                    .build_channel_proofs(&port_id, &channel_id, height)
-                    .map_err(BaseError::relayer)?;
-
-                let payload = CosmosChannelOpenConfirmPayload { proofs };
-
-                Ok(payload)
-            })
-            .await
-            .map_err(BaseError::join)?
+    ) -> Result<CosmosChannelOpenConfirmPayload, Error> {
+        build_channel_open_confirm_payload(self, height, port_id, channel_id).await
     }
 
     async fn build_channel_open_init_message(
@@ -1155,39 +1069,7 @@ where
         counterparty_port_id: &PortId,
         init_channel_options: &CosmosInitChannelOptions,
     ) -> Result<Arc<dyn CosmosMessage>, Error> {
-        let port_id = port_id.clone();
-        let ordering = init_channel_options.ordering;
-        let connection_hops = init_channel_options.connection_hops.clone();
-        let channel_version = init_channel_options.channel_version.clone();
-
-        let counterparty = ChannelCounterparty::new(counterparty_port_id.clone(), None);
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let channel = ChannelEnd::new(
-                    State::Init,
-                    ordering,
-                    counterparty,
-                    connection_hops,
-                    channel_version,
-                );
-
-                let message = CosmosIbcMessage::new(None, move |signer| {
-                    let message = MsgChannelOpenInit {
-                        port_id: port_id.clone(),
-                        channel: channel.clone(),
-                        signer: signer.clone(),
-                    };
-
-                    Ok(message.to_any())
-                });
-
-                Ok(wrap_cosmos_message(message))
-            })
-            .await
-            .map_err(BaseError::join)?
+        build_channel_open_init_message(port_id, counterparty_port_id, init_channel_options)
     }
 
     async fn build_channel_open_try_message(
@@ -1197,46 +1079,12 @@ where
         counterparty_channel_id: &ChannelId,
         counterparty_payload: CosmosChannelOpenTryPayload,
     ) -> Result<Arc<dyn CosmosMessage>, Error> {
-        let port_id = port_id.clone();
-        let counterparty = ChannelCounterparty::new(
-            counterparty_port_id.clone(),
-            Some(counterparty_channel_id.clone()),
-        );
-        let ordering = counterparty_payload.ordering;
-        let connection_hops = counterparty_payload.connection_hops.clone();
-        let version = counterparty_payload.version.clone();
-        let previous_channel_id = counterparty_payload.previous_channel_id.clone();
-        let proofs = counterparty_payload.proofs.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let channel = ChannelEnd::new(
-                    State::TryOpen,
-                    ordering,
-                    counterparty,
-                    connection_hops,
-                    version.clone(),
-                );
-
-                let message = CosmosIbcMessage::new(None, move |signer| {
-                    let message = MsgChannelOpenTry {
-                        port_id: port_id.clone(),
-                        previous_channel_id: previous_channel_id.clone(),
-                        channel: channel.clone(),
-                        counterparty_version: version.clone(),
-                        proofs: proofs.clone(),
-                        signer: signer.clone(),
-                    };
-
-                    Ok(message.to_any())
-                });
-
-                Ok(wrap_cosmos_message(message))
-            })
-            .await
-            .map_err(BaseError::join)?
+        build_channel_open_try_message(
+            port_id,
+            counterparty_port_id,
+            counterparty_channel_id,
+            counterparty_payload,
+        )
     }
 
     async fn build_channel_open_ack_message(
@@ -1246,33 +1094,12 @@ where
         counterparty_channel_id: &ChannelId,
         counterparty_payload: CosmosChannelOpenAckPayload,
     ) -> Result<Arc<dyn CosmosMessage>, Error> {
-        let port_id = port_id.clone();
-        let channel_id = channel_id.clone();
-        let counterparty_channel_id = counterparty_channel_id.clone();
-        let counterparty_version = counterparty_payload.version.clone();
-        let proofs = counterparty_payload.proofs.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let message = CosmosIbcMessage::new(None, move |signer| {
-                    let message = MsgChannelOpenAck {
-                        port_id: port_id.clone(),
-                        channel_id: channel_id.clone(),
-                        counterparty_channel_id: counterparty_channel_id.clone(),
-                        counterparty_version: counterparty_version.clone(),
-                        proofs: proofs.clone(),
-                        signer: signer.clone(),
-                    };
-
-                    Ok(message.to_any())
-                });
-
-                Ok(wrap_cosmos_message(message))
-            })
-            .await
-            .map_err(BaseError::join)?
+        build_channel_open_ack_message(
+            port_id,
+            channel_id,
+            counterparty_channel_id,
+            counterparty_payload,
+        )
     }
 
     async fn build_channel_open_confirm_message(
@@ -1281,28 +1108,6 @@ where
         channel_id: &ChannelId,
         counterparty_payload: CosmosChannelOpenConfirmPayload,
     ) -> Result<Arc<dyn CosmosMessage>, Error> {
-        let port_id = port_id.clone();
-        let channel_id = channel_id.clone();
-        let proofs = counterparty_payload.proofs.clone();
-
-        self.runtime
-            .runtime
-            .runtime
-            .spawn_blocking(move || {
-                let message = CosmosIbcMessage::new(None, move |signer| {
-                    let message = MsgChannelOpenConfirm {
-                        port_id: port_id.clone(),
-                        channel_id: channel_id.clone(),
-                        proofs: proofs.clone(),
-                        signer: signer.clone(),
-                    };
-
-                    Ok(message.to_any())
-                });
-
-                Ok(wrap_cosmos_message(message))
-            })
-            .await
-            .map_err(BaseError::join)?
+        build_channel_open_confirm_message(port_id, channel_id, counterparty_payload)
     }
 }
