+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,2 +1,3 @@
+pub mod channel;
 pub mod connection;
 pub mod update_client;
